roles/was/files/app/requirements.txt
fastapi
uvicorn
requests
pyjwt
python-multipart

sqlalchemy
psycopg2-binary
httpx